bot/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
cryptography==2.1.4
future==0.16.0
idna==2.6
peewee==3.1.3
Pillow==5.0.0
pycparser==2.18
python-telegram-bot==10.0.1
pyzbar==0.1.4
psycopg2==2.7.4
qrcode==5.3
six==1.11.0
